# Faster tree builder