Reintroduces one of the optimisations mentioned in https://github.com/google/xi-editor/issues/583.
This optimisation uses a height sorted stack to push a string on an existing tree.

I also added some benches. Benches with 'stack' in the name are with the new method:

`test build_tree_few_big_lines           ... bench:      90,491 ns/iter (+/- 6,478)`
`test build_tree_few_big_lines_stack ... bench:      80,863 ns/iter (+/- 4,891)`

`test build_tree_short_lines                     ... bench:     322,070 ns/iter (+/- 21,143)`
`test build_tree_short_lines_stack        	... bench:     209,952 ns/iter (+/- 9,921)`

`test build_tree_triangle                           ... bench:     503,447 ns/iter (+/- 36,269)`
`test build_tree_triangle_stack             	... bench:     299,227 ns/iter (+/- 23,092)`