@cmyr 2018-07-26
Cool, those are significant gains. Probably won't get around to giving this a thorough review for a few days, but happy to have you plugging away at perf stuff!