@mqzry 2018-07-26
I feel like there should be some tests.

I will add on sunday a test to see if the new method yields exectly the same rope as the old method.

Let me know if there are any other scenarios I should.