@mqzry 2018-08-27
Ready for next review or merge