@mqzry 2018-09-11
@raphlinus, this needs to be approved again. Incorporated the small nits of cmyr.