@cmyr 2018-09-11
Awesome, thanks for the comments and sorry for the delay. For what it's worth I've been spending a lot more time in the rope crate recently, I should be better able to review changes going forward.